requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
sentence-transformers
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
wikipedia
arxiv
langchain-groq
PyPDF2
docx2txt
